renamed login method to getCredentials
